View as structured array in `unique_rows`
To ensure the type information is preserved and accessible to NumPy,
view the array as a structured array. This pushes all but the 0th
dimension of the shape into the structured type. Thus leaving the 0th
dimension as the shape of the array. With this done, `unique`
effectively runs over the 0th axis.